Make oom_score Int32 to match the server setting; fix analyzer null-deref
The settings style check requires a setting name that exists in both Settings
and ServerSettings to have the same type. ServerSettings already declares
oom_score as Int32 (Linux OOM killer adjustment for the process), so the query
setting must be Int32 too. Int32 is more than enough for an eviction priority.

Also seed the victim search in selectAllocationToKill with the first running
allocation instead of a null pointer, so the clang static analyzer can see the
dereference is never null (the queue is already checked non-empty above).

+++ src/Common/Scheduler/MemoryReservation.cpp
@@ -32,7 +32,7 @@ namespace ErrorCodes
     extern const int MEMORY_RESERVATION_FAILED;
 }
 
-MemoryReservation::MemoryReservation(ResourceLink link, const String & id_, ResourceCost reserved_size_, Int64 oom_score_)
+MemoryReservation::MemoryReservation(ResourceLink link, const String & id_, ResourceCost reserved_size_, Int32 oom_score_)
     : ResourceAllocation(*link.allocation_queue, id_, oom_score_)
     , reserved_size(reserved_size_)
     , approved_increment(CurrentMetrics::MemoryReservationApproved, 0)
